requirements.txt
fastapi==0.104.1 # if you plan to add an API later
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
playwright==1.40.0
cairosvg==2.7.1 # for SVG→PNG conversion
gTTS==2.5.3 # simple TTS MVP (or use edge-tts)
moviepy==1.0.3
requests==2.32.3
httpx==0.25.2
pytest==7.4.3